Finished refactoring of firebase functions
- Now all the functions are in their corresponding js file.
- Fixed some issues and small problems with the update of the user info.
- Added extra validations to ChangeUserInfo.
- Deleted some console.log
