+++ LocalFoodApp/src/Components/GoogleBtn/GoogleBtn.js
@@ -4,7 +4,7 @@ import { GBtn, GBtnText } from './GoogleBtnStyles';
 import { GoogleSignin } from '@react-native-google-signin/google-signin'
 import auth from '@react-native-firebase/auth'
 import { webClientIdT3 } from '../../Others/AuthKeys/GoogleAuthKeys'
-import { NewUserDoc } from '../../Others/FirebaseFunctions/FirebaseFunctions'
+import { NewUserDoc } from '../../Others/FirebaseFunctions/UserFunctions'
 
 GoogleSignin.configure({
   webClientId: webClientIdT3,
@@ -17,7 +17,8 @@ export const GoogleBtn = () => {
         const googleCredential = auth.GoogleAuthProvider.credential(idToken);
         await auth().signInWithCredential(googleCredential).then((e)=>{
             if(e.additionalUserInfo.isNewUser) {
-                NewUserDoc(e)
+                // NewUserDoc(e)
+                console.log(e)
                 navigation.navigate('Home');
             }else {
                 navigation.navigate('Home');
