+++ LocalFoodApp/src/Others/ImagePicker/ImagePickerFunctions.js
@@ -15,7 +15,6 @@ export const requestCameraPermission = async () => {
         // If CAMERA Permission is granted
         return granted === PermissionsAndroid.RESULTS.GRANTED;
       } catch (err) {
-        console.warn(err);
         return false;
       }
     } else return true;
@@ -30,8 +29,6 @@ export const chooseFile = (setFilePath, setOpc) => {
         quality: 1,
     };
     launchImageLibrary(options, (response) => {
-        console.log('Response = ', response.assets[0]);
-
         if (response.didCancel) {
             alert('User cancelled camera picker');
             return;
