--- LocalFoodApp/src/Screens/Shop/Shop.js
@@ -1,12 +1,0 @@
-import { Text, View } from 'react-native';
-import React from 'react';
-
-export const Shop = (props) => {
-  const { route: { params: { shop } } } = props
-  const { ShopName } = shop
-  return (
-    <View>
-      <Text>{ShopName}</Text>
-    </View>
-  );
-};
