+++ LocalFoodApp/src/Screens/UserPanel/UserPanel.js
@@ -6,7 +6,7 @@ import { styles } from './UserPStyles';
 import auth from '@react-native-firebase/auth'
 import { MainBtn } from '../../Components/MainBtn/MainBtn';
 import { UserInfo } from '../../Components/UserInfo/UserInfo';
-import { UserOwnerInfo } from '../../Others/FirebaseFunctions/FirebaseFunctions';
+import { UserOwnerInfo } from '../../Others/FirebaseFunctions/UserFunctions';
 
 export const UserPanel = ({navigation}) => {
   const [userIsOwner, setUserIsOwner] = useState(null);
